+++ src/FileHelper.php
@@ -440,7 +440,7 @@ class FileHelper
      * @param bool $octal default false i.e. return '644'
      * @return bool|string return false if it fails.
      */
-    public static function file_perms(string $file, bool $octal = false)
+    public static function filPerms(string $file, bool $octal = false)
     {
         if (!file_exists($file)) {
             return false;